// frontend/src/api/generated/models/productVariantPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { ProductVariantPayloadMetadata } from './productVariantPayloadMetadata.ts';

export interface ProductVariantPayload {
  product_model_id: string;
  color?: string;
  size?: string;
  metadata?: ProductVariantPayloadMetadata;
  default_margin_percent?: number | string | null;
  is_active?: boolean;
}