// frontend/src/api/generated/models/productVariantResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { ProductVariantResponseMetadata } from './productVariantResponseMetadata.ts';

export interface ProductVariantResponse {
  id: string;
  product_model_id: string;
  sku: string;
  color: string;
  size: string;
  metadata: ProductVariantResponseMetadata;
  default_margin_percent: string | null;
  is_active: boolean;
}